merge

+++ configcenter-biz/src/main/java/org/antframework/configcenter/biz/service/DeleteProfileService.java
@@ -8,19 +8,18 @@
  */
 package org.antframework.configcenter.biz.service;
 
+import org.antframework.common.util.facade.BizException;
+import org.antframework.common.util.facade.CommonResultCode;
 import org.antframework.common.util.facade.EmptyResult;
-import org.antframework.common.util.facade.FacadeUtils;
+import org.antframework.common.util.facade.Status;
 import org.antframework.configcenter.biz.util.AppUtils;
-import org.antframework.configcenter.biz.util.PropertyKeyUtils;
+import org.antframework.configcenter.biz.util.PropertyValueUtils;
 import org.antframework.configcenter.biz.util.RefreshUtils;
+import org.antframework.configcenter.biz.util.ReleaseUtils;
 import org.antframework.configcenter.dal.dao.ProfileDao;
 import org.antframework.configcenter.dal.entity.Profile;
-import org.antframework.configcenter.facade.api.PropertyValueService;
 import org.antframework.configcenter.facade.info.AppInfo;
-import org.antframework.configcenter.facade.info.PropertyKeyInfo;
 import org.antframework.configcenter.facade.order.DeleteProfileOrder;
-import org.antframework.configcenter.facade.order.SetPropertyValuesOrder;
-import org.antframework.configcenter.facade.vo.Scope;
 import org.bekit.service.annotation.service.Service;
 import org.bekit.service.annotation.service.ServiceAfter;
 import org.bekit.service.annotation.service.ServiceExecute;
@@ -34,8 +33,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 public class DeleteProfileService {
     @Autowired
     private ProfileDao profileDao;
-    @Autowired
-    private PropertyValueService propertyValueService;
 
     @ServiceExecute
     public void execute(ServiceContext<DeleteProfileOrder, EmptyResult> context) {
@@ -45,33 +42,16 @@ public class DeleteProfileService {
         if (profile == null) {
             return;
         }
-        // 删除所有应用在该环境下的所有属性value
+        if (profileDao.existsByParent(order.getProfileId())) {
+            throw new BizException(Status.FAIL, CommonResultCode.ILLEGAL_STATE.getCode(), String.format("环境[%s]存在子环境，不能删除", order.getProfileId()));
+        }
+        // 删除所有应用在该环境下的配置value和发布
         for (AppInfo app : AppUtils.findAllApps()) {
-            deleteAppProfileAllPropertyValues(app.getAppId(), order.getProfileId());
+            PropertyValueUtils.deleteAppProfilePropertyValues(app.getAppId(), order.getProfileId());
+            ReleaseUtils.deleteAppProfileReleases(app.getAppId(), order.getProfileId());
         }
         // 删除环境
         profileDao.delete(profile);
-    }
-
-    // 删除应有在指定环境的所有属性value
-    private void deleteAppProfileAllPropertyValues(String appId, String profileId) {
-        SetPropertyValuesOrder order = new SetPropertyValuesOrder();
-        order.setAppId(appId);
-        order.setProfileId(profileId);
-
-        for (PropertyKeyInfo propertyKey : PropertyKeyUtils.findAppPropertyKeys(appId, Scope.PRIVATE)) {
-            SetPropertyValuesOrder.KeyValue keyValue = new SetPropertyValuesOrder.KeyValue();
-            keyValue.setKey(propertyKey.getKey());
-            keyValue.setValue(null);
-
-            order.addKeyValue(keyValue);
-        }
-        if (order.getKeyValues().size() <= 0) {
-            return;
-        }
-
-        EmptyResult result = propertyValueService.setPropertyValues(order);
-        FacadeUtils.assertSuccess(result);
     }
 
     @ServiceAfter
